requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.1.0
babel==2.17.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
daff==1.3.46
db-dtypes==1.4.1
dbt-adapters==1.13.2
dbt-bigquery==1.9.1
dbt-common==1.14.0
dbt-core==1.9.2
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.16.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.6.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.2
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
snowplow-tracker==1.0.4
sqlparse==0.5.3
text-unidecode==1.3
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
